+++ com.unity.mobile.android-logcat/Tests/Editor/AndroidLogcatPackageTests.cs
@@ -20,20 +20,20 @@ internal class AndroidLogcatPackageTests : AndroidLogcatRuntimeTestBase
             var fakeDevice = new AndroidLogcatFakeDevice90("androiddevice0");
             for (int i = 0; i < kPackagesToCreate; i++)
             {
-                var d = m_Runtime.ProjectSettings.CreatePackageInformation("com.unity.test" + i, i + 1, fakeDevice);
+                var d = m_Runtime.UserSettings.CreatePackageInformation("com.unity.test" + i, i + 1, fakeDevice);
             }
 
             // All packages are alive, calling cleanup dead packages, shouldn't clean anything
-            m_Runtime.ProjectSettings.CleanupDeadPackagesForDevice(fakeDevice);
-            var packages = m_Runtime.ProjectSettings.GetKnownPackages(fakeDevice);
+            m_Runtime.UserSettings.CleanupDeadPackagesForDevice(fakeDevice);
+            var packages = m_Runtime.UserSettings.GetKnownPackages(fakeDevice);
             Assert.AreEqual(kPackagesToCreate, packages.Count);
 
             foreach (var p in packages)
                 p.SetExited();
 
-            m_Runtime.ProjectSettings.CleanupDeadPackagesForDevice(fakeDevice);
+            m_Runtime.UserSettings.CleanupDeadPackagesForDevice(fakeDevice);
 
-            Assert.AreEqual(AndroidLogcatProjectSettings.kMaxExitedPackages, packages.Count);
+            Assert.AreEqual(AndroidLogcatUserSettings.kMaxExitedPackages, packages.Count);
 
             // Check that recent packages are still there, only the old packages should be removed
             foreach (var p in packages)