+++ com.unity.mobile.android-logcat/Tests/Editor/AndroidLogcatTestRuntime.cs
@@ -5,9 +5,9 @@ using Unity.Android.Logcat;
 
 internal class AndroidLogcatTestRuntime : AndroidLogcatRuntimeBase
 {
-    internal static readonly string kProjectSettingsPath = Path.Combine("Tests", "ProjectSettings", "AndroidLogcatSettings.asset");
+    internal static readonly string kUserSettingsPath = Path.Combine("Tests", "UserSettings", "AndroidLogcatSettings.asset");
 
-    protected override string ProjectSettingsPath { get => kProjectSettingsPath; }
+    protected override string UserSettingsPath { get => kUserSettingsPath; }
 
     public override AndroidLogcatMessageProviderBase CreateMessageProvider(AndroidBridge.ADB adb, string filter, AndroidLogcat.Priority priority, int packageID, string logPrintFormat, IAndroidLogcatDevice device, Action<string> logCallbackAction)
     {